Move capture and cssOptimize to separate repos
