+++ tests/index.html
@@ -97,10 +97,7 @@
     });
     
 </script>
-  <script src="capture-tests.js"></script>
   <script src="unblockify-tests.js"></script>
-  <script src="cssOptimize-tests.js"></script>
-  <script src="anchor-tests.js"></script>
 
   <script src="/tag/supportedBrowser.js"></script>
   <script src="supportedBrowser-tests.js"></script>
